+++ internal/token/base62.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Generator interface {
-	GetToken(id int) string
+	GetToken(id uint64) string
 }
 
 type base64TokenGenerator struct {
@@ -15,8 +15,8 @@ type base64TokenGenerator struct {
 
 // GetToken divides the input integer ID into segments, each with a maximum length of 10, as per the Max base62 token length.
 // in response; returns the result as a concatenated string of generated tokens for each segment.
-func (tg *base64TokenGenerator) GetToken(id int) string {
-	strID := strconv.Itoa(id)
+func (tg *base64TokenGenerator) GetToken(id uint64) string {
+	strID := strconv.Itoa(int(id))
 	var result strings.Builder
 
 	for len(strID) > 0 {
